+++ Test/error_test.py
@@ -48,6 +48,10 @@ pycm.pycm_obj.pycmMatrixError: Mapping format error
 Traceback (most recent call last):
         ...
 pycm.pycm_obj.pycmMatrixError: Mapping class names error
+>>> cm.relabel({0:"L1",1:"L2",2:"L2"})
+Traceback (most recent call last):
+        ...
+pycm.pycm_obj.pycmMatrixError: Mapping class names error
 >>> cp = Compare([cm,cm])
 Traceback (most recent call last):
         ...